style: :lipstick: 내비게이션 바 컴포넌트 분할, 375px 미만일 때 디자인 수정
